% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFL_prod_tryCatchLog.R
\name{rfl_logRunError}
\alias{rfl_logRunError}
\title{Log Errors}
\usage{
rfl_logRunError(out, ...)
}
\arguments{
\item{out}{A string or R error, the thing you want to log.}

\item{RProcessName}{A string, the name of the RFL proccess the function is being run.}

\item{RScriptName}{A string, the script file being run.}

\item{RScriptLocation}{A string, where the file is stored}

\item{RErrorLogName}{A string, the log of the run}

\item{RErrorLogLocation}{A string, the log location}
}
\value{
Success if no errors, or failure and failure logged in the data warehouse
}
\description{
Log Errors
}
\examples{
rfl_logRunError(
RProcessName = "PerformancePulse",
RScriptName = "deploy_app.R",
RScriptLocation = "\\\\\\\\rfh-information\\\\G$\\\\RScripts\\\\PerformancePulse",
RErrorLogName	=	"deploy_app_outfile.ROut",
RErrorLogLocation = "\\\\\\\\rfh-information\\\\G$\\\\RScripts\\\\PerformancePulse\\\\logs"
)
}
