% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFL_prod_tryCatchLog.R
\name{rfl_tryCatchLog}
\alias{rfl_tryCatchLog}
\title{Try to Run A Function, log if There's an Error}
\usage{
rfl_tryCatchLog(fun, RProcessName, RScriptName, RScriptLocation,
  RErrorLogName, RErrorLogLocation)
}
\arguments{
\item{fun}{A function to run,}

\item{RProcessName}{A string, the name of the RFL proccess the function is being run.}

\item{RScriptName}{A string, the script file being run.}

\item{RScriptLocation}{A string, where the file is stored}

\item{RErrorLogName}{A string, the log of the run}

\item{RErrorLogLocation}{A string, the log location}
}
\value{
Success if no errors, or failure and failure logged in the data warehouse
}
\description{
Try to Run A Function, log if There's an Error
}
\examples{
rfl_tryCatchLog(
sum(c(1,2)),
RProcessName = "PerformancePulse",
RScriptName = "deploy_app.R",
RScriptLocation = "\\\\\\\\rfh-information\\\\G$\\\\RScripts\\\\PerformancePulse",
RErrorLogName	=	"deploy_app_outfile.ROut",
RErrorLogLocation = "\\\\\\\\rfh-information\\\\G$\\\\RScripts\\\\PerformancePulse\\\\logs"
)
}
